Mini-Project-1/blog_api/requirements.txt
Django>=4.0
djangorestframework
django-cors-headers
Pillow
python-dotenv